refactor pre-processing DB code to its own folder
